pyproject.toml
[project]
name = "esopn"
version = "0.1.0"
description = "AI Sports Commentator Duo - Real-time commentary for AI coding agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mss>=10.0.0",
    "google-genai>=1.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.52.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pynput>=1.7.0",
    "elevenlabs>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
esopn = "esopn.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/esopn"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
